+++ src/Request/Event.php
@@ -114,7 +114,7 @@ class Event extends RequestCollection
             case 'feed_timeline':
             case 'explore_popular':
             case 'newsfeed_you':
-                $class = '1IY';
+                $class = '1Wn';
                 break;
             case 'search':
             case 'blended_search':
@@ -122,23 +122,23 @@ class Event extends RequestCollection
             case 'search_tags':
             case 'search_places':
             case 'search_result':
-                $class = 'A82';
+                $class = '9YJ';
                 break;
             case 'feed_hashtag':
-                $class = '8Se';
+                $class = '8v8';
                 break;
             case 'feed_location':
-                $class = '8RA';
+                $class = '8vO';
                 break;
             case 'feed_contextual_chain':
-                $class = '8Xn';
+                $class = '8gH';
                 break;
             case 'feed_contextual_place':
             case 'feed_contextual_location':
             case 'feed_contextual_hashtag':
             case 'feed_contextual_profile':
             case 'feed_contextual_self_profile':
-                $class = '8Mk';
+                $class = '8gx';
                 break;
             case 'profile':
             case 'self_profile':
@@ -152,13 +152,13 @@ class Event extends RequestCollection
                 $class = 'UnifiedFollowFragment';
                 break;
             case 'likers':
-                $class = '7u1';
+                $class = '6jI';
                 break;
             case 'tabbed_gallery_camera':
                 $class = 'MediaCaptureFragment';
                 break;
             case 'photo_filter':
-                $class = 'Ds4';
+                $class = 'Ajm';
                 break;
             case 'metadata_followers_share':
                 $class = 'FollowersShareFragment';